requirements.txt
# Core data & ML
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
scipy>=1.11.0

# Deep learning
torch>=2.0.0

# Collaborative filtering
implicit>=0.7.2

# Approximate nearest neighbor search
faiss-cpu>=1.7.4

# API serving
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0

# Data download & utilities
requests>=2.31.0
tqdm>=4.66.0
PyYAML>=6.0

# Visualization & notebooks
matplotlib>=3.7.0
seaborn>=0.13.0
jupyterlab>=4.0.0

# Statistical testing
scipy>=1.11.0
